Fixed the CRC algorithm (although something is still not quite right
- I also tried to get Preon to work without any luck. It seems that it is parsing the fields in reverse order.
- Started adding a new theme, but have not used it yet.

+++ app/src/main/java/com/craxiom/networksurveyplus/messages/QcdmMessage.java
@@ -18,7 +18,7 @@ import java.util.Arrays;
  */
 public class QcdmMessage
 {
-    public static final byte[] QCDM_PREFIX = {(byte) 0x98, 0x01, 0x00, 0x00, 0x01, 0x00, 0x00, 0x00};
+    public static final byte[] QCDM_PREFIX = {(byte) 0x98, (byte) 0x01, (byte) 0x00, (byte) 0x00, (byte) 0x01, (byte) 0x00, (byte) 0x00, (byte) 0x00};
     public static final int QCDM_FOOTER = 0x7e;
 
     /**
